Replace std::map with base::SmallMap in ui::LatencyInfo
The typical number of components per LatencyInfo instance is quite small. Avoid
unnecessary heap allocation by using a SmallMap to store component entries.
This reduces the average browser send time cost per touch or scroll event by
~15us on a Nexus 4, and ~22us on a Galaxy Nexus.

BUG=341613

Review URL: https://codereview.chromium.org/157003005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257740 0039d316-1c4b-4281-b951-d872f2087c98

+++ ui/events/latency_info.cc
@@ -13,7 +13,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace {
 
-const unsigned int kMaxLatencyInfoNumber = 100;
+const size_t kMaxLatencyInfoNumber = 100;
 
 const char* GetComponentName(ui::LatencyComponentType type) {
 #define CASE_TYPE(t) case ui::t:  return #t
